+++ bitalarm-impl/src/test/java/com/dwarfeng/bitalarm/impl/service/AlarmHistoryMaintainServiceImplTest.java
@@ -60,7 +60,7 @@ public class AlarmHistoryMaintainServiceImplTest {
             AlarmHistory alarmHistory = new AlarmHistory(
                     null,
                     null,
-                    12450,
+                    null,
                     1,
                     "我是报警信息",
                     (byte) 0,